Always free XDR-allocated buffers
In a few places, we decode an XDR stream that results in XDR-allocated
buffers (either via an RPC call or a manual xdr_* call), but we don't
try to free those resources if the decoding failed, or in similar edge
cases. If decoding an XDR stream fails, memory may still have been
allocated, and so we must always try to free the relevant resources.
Change the relevant code to always run xdr_free or similar after
decoding.

Change-Id: I559d4f7c8493338cf09168a5133f89f966af204c
Reviewed-on: https://gerrit.openafs.org/16237
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>