xdr: Blank non-preallocated OUT arguments
When calling an RPC with dynamically-allocated OUT arguments, the
caller needs to either blank the memory given to the RPC, or provide a
preallocated buffer. A few callers were just providing uninitialized
memory, which can look like a preallocated buffer that points to bogus
memory.

Fix these callers to blank the relevant memory, so we always get
XDR-allocated memory back. All of these callers are internal/debug tools
that shouldn't see much use, but fix them in case they ever are used.

Change-Id: I705db6bbb9c6afd97b1a7659dda241aa5cca2269
Reviewed-on: https://gerrit.openafs.org/16241
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>