pts: Use pr_NameToId in pr_SNameToId
Currently, pr_SNameToId does a couple of things that pr_NameToId also
does internally: run stolower() on the name, and check that we didn't
get back an empty array of ids. Change pr_SNameToId to just run
pr_NameToId instead, to avoid duplicating this logic and slightly
simplify the function.

Change-Id: Ic587cd2b37ae4015d6ac7f7166dd6216650e6a70
Reviewed-on: https://gerrit.openafs.org/16239
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>