vlserver: Add audit events for all RPCs
Currently, most RPCs in the vlserver generate an audit event, but not
all: notably, the most common operations, VL_GetEntryBy*, do not
generate audit events.

To be able to track all vlserver activity in the audit log, create new
audit events for all of the missing VL RPCs, and generate those audit
events from the relevant RPCs.

Also introduce a couple of new audit data types, AUD_BULKADDRS and
AUD_AFSUUID, in order to record the arguments for some of the
newly-audited RPCs.

Change-Id: Ifd889479f9a0c124a44b8015c907f014c8e1f215
Reviewed-on: https://gerrit.openafs.org/14467
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>